% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{error}
\alias{error}
\title{Compute Community Assignment Error}
\usage{
error(n, c, es_c)
}
\arguments{
\item{n}{the length of the input community vector}

\item{c}{the true community assignment vector}

\item{es_c}{the estimated community assignment vector}
}
\value{
the computational error between two input vectors
}
\description{
Calculate the error between the true community vector and the estimated community vector
}
