% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_detection.R
\name{main}
\alias{main}
\title{Community Detection using the Improved EM Algorithm in R}
\usage{
main(real)
}
\arguments{
\item{real}{the input matrix of data}
}
\value{
a vector of the estimated community assignments
}
\description{
This function repeatedly estimates the community assignments by applying the Improved EM algorithm to the matrix parameters
}
