% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_detection.R
\name{real_ac}
\alias{real_ac}
\title{Normal SBM Matrix Generator using R}
\usage{
real_ac(dm, dv)
}
\arguments{
\item{dm}{the parameter that determines the normal distribution mean for generating the adjacency matrix}

\item{dv}{the parameter that determines the normal distribution variance for generating the adjacency matrix}
}
\value{
A list containing:
\item{alpha}{The probability of community assignments for each node.}
\item{mu}{The true mean of the normal distribution used to generate the matrix.}
\item{sigma}{The true variance of the normal distribution used to generate the matrix.}
\item{c}{A vector of community assignments for the nodes.}
\item{A}{A generated normal weighted stochastic block matrix.}
}
\description{
Simulating Normal Weighted Stochastic Block Model Matrices using R
}
