\name{StatComp-package}
\alias{StatComp-package}
\docType{package}
\title{
  Community Detection on Weighted Networks
}
\description{
  R and Rcpp versions of weighted stochastic block model and variational EM algorithm are provided in this package (\code{wsbmR} and \code{vemR} versus \code{wsbmC} and \code{vemC}). These functions allow users to estimate model parameters and obtain community structures of weighted networks.
}

\author{
Lei Liu

Maintainer: Lei Liu <ll0726@mail.ustc.edu.cn>
}
\keyword{ package }
